Improved handling of inconsistent data, added a flag to trends, general refactoring and tidy-up
Improved handling of 'problematic' data in METAR, SPECI and TAF reports: the following rules are established for contradictory, duplicate or inconsistent data which may potentially occur in the report.

 - Duplicate groups, groups providing the same data in different form and groups providing no new data are because they provide no additional information. No warning is added.
 - Groups which contradict each other are ignored (all of them) because there is no way to know which group to trust. Warning is added.

Added flag to trends which allows to distinguis between trends specified in METAR/SPECI report and in TAF report; this will be useful for merging a pair of METAR report and TAF report into a single structure.

Some tidy-up and refactoring of the code without functionality change.